+++ tests/testthat/test-objective_function.R
@@ -91,6 +91,24 @@ test_that('Objective functions can be created and behave as expected', {
         )
     )
 
+    # Make sure there are no errors for one data-driver pair, no dependent
+    # arguments in verbose mode
+    sink(tempfile())
+
+    expect_no_error(
+        objective_function(
+            model,
+            ddps[1],
+            independent_args,
+            quantity_weights,
+            data_definitions = data_definitions,
+            post_process_function = post_process_function,
+            verbose_startup = TRUE
+        )
+    )
+
+    sink()
+
     # Two data-driver pairs, with dependent arguments and L2 regularization
     obj_fun <- expect_silent(
         objective_function(
@@ -282,6 +300,159 @@ test_that('Data-driver pairs must have correct elements', {
             verbose_startup = verbose_startup
         ),
         'The following data-driver pairs have a `data_stdev` element that does not match the columns and/or times of their `data` element: ambient_2002'
+    )
+})
+
+test_that('Driver-specific initial values are checked and used', {
+    expect_error(
+        objective_function(
+            model,
+            within(ddps, {ambient_2002$initial_values = 5}),
+            independent_args,
+            quantity_weights,
+            data_definitions = data_definitions,
+            post_process_function = post_process_function,
+            verbose_startup = verbose_startup
+        ),
+        'When provided, the driver-specific initial values must be a list of named elements'
+    )
+
+    expect_error(
+        objective_function(
+            model,
+            within(ddps, {ambient_2002$initial_values = list(5)}),
+            independent_args,
+            quantity_weights,
+            data_definitions = data_definitions,
+            post_process_function = post_process_function,
+            verbose_startup = verbose_startup
+        ),
+        'When provided, the driver-specific initial values must be a list of named elements'
+    )
+
+    expect_error(
+        objective_function(
+            model,
+            within(ddps, {ambient_2002$initial_values = list(Leaf = 5)}),
+            independent_args,
+            quantity_weights,
+            data_definitions = data_definitions,
+            post_process_function = post_process_function,
+            verbose_startup = verbose_startup
+        ),
+        'The following driver-specific initial value names were provided in the data-driver pairs:\nambient_2002 : Leaf\nambient_2005 : \nWhen provided, these names must be the same for each set of drivers',
+        fixed = TRUE
+    )
+
+    expect_error(
+        objective_function(
+            model,
+            within(ddps, {
+                ambient_2002$initial_values = list(extra_iv = 5)
+                ambient_2005$initial_values = list(extra_iv = 7)
+            }),
+            independent_args,
+            quantity_weights,
+            data_definitions = data_definitions,
+            post_process_function = post_process_function,
+            verbose_startup = verbose_startup
+        ),
+        'The following driver-specific initial values are not included in the base model definition: "extra_iv"'
+    )
+
+    # Here we specify a "bad" initial value of Leaf for 2005. This will trigger
+    # an error, but the particular error may depend on the particular version of
+    # BioCro being used, so we don't check the error message content.
+    expect_error(
+        objective_function(
+            model,
+            within(ddps, {
+                ambient_2002$initial_values = list(Leaf = 0.1) # This shouldn't cause a problem
+                ambient_2005$initial_values = list(Leaf = -1)  # This should trigger an error
+            }),
+            independent_args,
+            quantity_weights,
+            data_definitions = data_definitions,
+            post_process_function = post_process_function,
+            verbose_startup = verbose_startup
+        )
+    )
+})
+
+test_that('Driver-specific parameters are checked and used', {
+    expect_error(
+        objective_function(
+            model,
+            within(ddps, {ambient_2002$parameters = 5}),
+            independent_args,
+            quantity_weights,
+            data_definitions = data_definitions,
+            post_process_function = post_process_function,
+            verbose_startup = verbose_startup
+        ),
+        'When provided, the driver-specific parameters must be a list of named elements'
+    )
+
+    expect_error(
+        objective_function(
+            model,
+            within(ddps, {ambient_2002$parameters = list(5)}),
+            independent_args,
+            quantity_weights,
+            data_definitions = data_definitions,
+            post_process_function = post_process_function,
+            verbose_startup = verbose_startup
+        ),
+        'When provided, the driver-specific parameters must be a list of named elements'
+    )
+
+    expect_error(
+        objective_function(
+            model,
+            within(ddps, {ambient_2002$parameters = list(Catm = 5)}),
+            independent_args,
+            quantity_weights,
+            data_definitions = data_definitions,
+            post_process_function = post_process_function,
+            verbose_startup = verbose_startup
+        ),
+        'The following driver-specific parameter names were provided in the data-driver pairs:\nambient_2002 : Catm\nambient_2005 : \nWhen provided, these names must be the same for each set of drivers',
+        fixed = TRUE
+    )
+
+    expect_error(
+        objective_function(
+            model,
+            within(ddps, {
+                ambient_2002$parameters = list(extra_param = 5)
+                ambient_2005$parameters = list(extra_param = 7)
+            }),
+            independent_args,
+            quantity_weights,
+            data_definitions = data_definitions,
+            post_process_function = post_process_function,
+            verbose_startup = verbose_startup
+        ),
+        'The following driver-specific parameters are not included in the base model definition: "extra_param"'
+    )
+
+    # Here we specify a "bad" value of the Ball-Berry intercept for 2005. This
+    # will trigger an error, but the particular error may depend on the
+    # particular version of BioCro being used, so we don't check the error
+    # message content.
+    expect_error(
+        objective_function(
+            model,
+            within(ddps, {
+                ambient_2002$parameters = list(b0 = 0.1) # This shouldn't cause a problem
+                ambient_2005$parameters = list(b0 = -1)  # This should trigger an error
+            }),
+            independent_args,
+            quantity_weights,
+            data_definitions = data_definitions,
+            post_process_function = post_process_function,
+            verbose_startup = verbose_startup
+        )
     )
 })